Tsify lib/shorteners (#4504)
* Convert lib/shorteners to ts

* Improved types
