+++ lib/shortener/_all.ts
