+++ lib/shortener/base.ts
@@ -1,4 +1,4 @@
-// Copyright (c) 2017, Compiler Explorer Authors
+// Copyright (c) 2018, Compiler Explorer Authors
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -22,29 +22,17 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-import request from 'request';
+import * as express from 'express';
 
-export class ShortLinkResolver {
-    resolve(url) {
-        return new Promise((resolve, reject) => {
-            request({method: 'HEAD', uri: url, followRedirect: false}, (err, res) => {
-                if (err !== null) {
-                    reject(err.message);
-                    return;
-                }
-                if (res.statusCode !== 302) {
-                    reject(`Got response ${res.statusCode}`);
-                    return;
-                }
-                const targetLocation = res.headers['location'];
-                if (!targetLocation) {
-                    reject(`Missing location url in ${targetLocation}`);
-                    return;
-                }
-                resolve({
-                    longUrl: targetLocation,
-                });
-            });
-        });
+import {StorageBase} from '../storage';
+
+export abstract class BaseShortener {
+    constructor(protected storageHandler: StorageBase) {}
+
+    // eslint-disable-next-line no-unused-vars
+    abstract handle(req: express.Request, res: express.Response);
+
+    static get key(): string {
+        throw 'get key() must be overridden';
     }
 }
