+++ lib/shortener/google.ts
@@ -1,4 +1,4 @@
-// Copyright (c) 2018, Compiler Explorer Authors
+// Copyright (c) 2017, Compiler Explorer Authors
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -22,14 +22,29 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-import {BaseShortener} from './base';
+import request from 'request';
 
-export class DefaultShortener extends BaseShortener {
-    static get key() {
-        return 'default';
-    }
-
-    handle(req, res) {
-        return this.storageHandler.handler(req, res);
+export class ShortLinkResolver {
+    resolve(url: string) {
+        return new Promise((resolve, reject) => {
+            request({method: 'HEAD', uri: url, followRedirect: false}, (err, res) => {
+                if (err !== null) {
+                    reject(err.message);
+                    return;
+                }
+                if (res.statusCode !== 302) {
+                    reject(`Got response ${res.statusCode}`);
+                    return;
+                }
+                const targetLocation = res.headers['location'];
+                if (!targetLocation) {
+                    reject(`Missing location url in ${targetLocation}`);
+                    return;
+                }
+                resolve({
+                    longUrl: targetLocation,
+                });
+            });
+        });
     }
 }
