+++ lib/shortener/index.ts
